fix: pool builder, migrate type, app async, and .gitignore
- shared/db: remove explicit wait_timeout from Pool::builder to avoid
  BuildError::TimeoutRequiresRuntime under tokio 1.53.x Handle::try_current
- shared/migrate: change execution_ms from i64 to i32 to match schema
- oa4rust/src: make program_center_core_entity router async; add global
  Extension<Pool> and Extension<DatabaseConnection> layers to create_app
- main: make ToolBridge::new async; apply same router fix in http path
- .gitignore: add rust target/, .pgdata/, *.log patterns
- delete stale feasibility_review.json and review-findings.json

+++ oa4rust/crates/shared/src/migrate.rs
@@ -129,7 +129,7 @@ pub async fn run_migrations(pool: &Pool) -> anyhow::Result<MigrationReport> {
         tx.batch_execute(&sql)
             .await
             .map_err(|e| anyhow::anyhow!("migration {} failed: {}", name, e))?;
-        let ms = start.elapsed().as_millis() as i64;
+        let ms = start.elapsed().as_millis() as i32;
         tx.execute(
             "INSERT INTO schema_migrations (version, checksum, execution_ms) VALUES ($1, $2, $3)",
             &[&name, &checksum, &ms],
